B038 Renamed helpers to helpers/index
